+++ packages/providers/lib.esm/alchemy-provider.js
@@ -1,4 +1,5 @@
 "use strict";
+import { WebSocketProvider } from "./websocket-provider";
 import { Logger } from "@ethersproject/logger";
 import { version } from "./_version";
 const logger = new Logger(version);
@@ -9,6 +10,12 @@ import { UrlJsonRpcProvider } from "./url-json-rpc-provider";
 //   https://dashboard.alchemyapi.io
 const defaultApiKey = "_gg7wSSi0KMBsdKnGVfHDueq6xMB9EkC";
 export class AlchemyProvider extends UrlJsonRpcProvider {
+    static getWebSocketProvider(network, apiKey) {
+        const provider = new AlchemyProvider(network, apiKey);
+        const url = provider.connection.url.replace(/^http/i, "ws")
+            .replace(".alchemyapi.", ".ws.alchemyapi.");
+        return new WebSocketProvider(url, provider.network);
+    }
     static getApiKey(apiKey) {
         if (apiKey == null) {
             return defaultApiKey;
@@ -22,19 +29,19 @@ export class AlchemyProvider extends UrlJsonRpcProvider {
         let host = null;
         switch (network.name) {
             case "homestead":
-                host = "eth-mainnet.alchemyapi.io/jsonrpc/";
+                host = "eth-mainnet.alchemyapi.io/v2/";
                 break;
             case "ropsten":
-                host = "eth-ropsten.alchemyapi.io/jsonrpc/";
+                host = "eth-ropsten.alchemyapi.io/v2/";
                 break;
             case "rinkeby":
-                host = "eth-rinkeby.alchemyapi.io/jsonrpc/";
+                host = "eth-rinkeby.alchemyapi.io/v2/";
                 break;
             case "goerli":
-                host = "eth-goerli.alchemyapi.io/jsonrpc/";
+                host = "eth-goerli.alchemyapi.io/v2/";
                 break;
             case "kovan":
-                host = "eth-kovan.alchemyapi.io/jsonrpc/";
+                host = "eth-kovan.alchemyapi.io/v2/";
                 break;
             default:
                 logger.throwArgumentError("unsupported network", "network", arguments[0]);