+++ packages/providers/lib/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "providers/5.0.3";
+export declare const version = "providers/5.0.4";